(new API) walk() accepts partial-application of callbacks!
This has been part of the purpose of giraphe since the beginning; I left
it out of the initial implementation, because I was using a hap-hazard
stateful algorithm. I'd *intended* to replace the stateful algorithm
with a relatively-pure recursive one (graph-walking seems like an ideal
place to actually *use* functional code, frankly) — unfortunately, after
struggling with that goal for a few hours, I kept running into nasty
design constraints that were annoying me.

So, this is finally a half-assed, but functional, implementation
thereof.

/squashes 575e6afe..c391d168